# docs/swagger/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost/",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/channels/{name}": {
            "get": {
                "tags": [
                    "channels-controller"
                ],
                "operationId": "getItem",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StoredPackageInfo"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/channels/{channel}/{name}/{version}": {
            "get": {
                "tags": [
                    "channels-controller"
                ],
                "operationId": "packageVersionInfo",
                "parameters": [
                    {
                        "name": "channel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoredVersionInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/channels/": {
            "get": {
                "tags": [
                    "channels-controller"
                ],
                "operationId": "list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChannelInfo"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/channels/{channel}/{name}": {
            "get": {
                "tags": [
                    "channels-controller"
                ],
                "operationId": "packageInfo",
                "parameters": [
                    {
                        "name": "channel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoredPackageInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/channels/{channel}/{name}/versions": {
            "get": {
                "tags": [
                    "channels-controller"
                ],
                "operationId": "versions",
                "parameters": [
                    {
                        "name": "channel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StoredVersionInfo"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/download/{name}/{filename}": {
            "get": {
                "tags": [
                    "compatibility-controller"
                ],
                "operationId": "downloadPackage",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filename",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/download/list.txt": {
            "get": {
                "tags": [
                    "compatibility-controller"
                ],
                "operationId": "packageListTXT",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/service/tasks/{name}/status": {
            "get": {
                "tags": [
                    "jobs-controller"
                ],
                "operationId": "getStatus",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobStatus"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/service/tasks/{name}/start": {
            "get": {
                "tags": [
                    "jobs-controller"
                ],
                "operationId": "startTask",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobStatus"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/service/tasks/{name}": {
            "get": {
                "tags": [
                    "jobs-controller"
                ],
                "operationId": "getJob",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobInfo_Public"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/packages": {
            "get": {
                "tags": [
                    "packages-controller"
                ],
                "operationId": "packageList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StoredPackageInfo"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/packages/{name}/{version}": {
            "get": {
                "tags": [
                    "packages-controller"
                ],
                "operationId": "packageVersionInfo_1",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoredVersionInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/push": {
            "post": {
                "tags": [
                    "packages-controller"
                ],
                "operationId": "pushPackage_1",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/packages/{name}": {
            "get": {
                "tags": [
                    "packages-controller"
                ],
                "operationId": "packageInfo_1",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoredPackageInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/packages/{name}/versions": {
            "get": {
                "tags": [
                    "packages-controller"
                ],
                "operationId": "versions_1",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StoredVersionInfo"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoredPackageInfo": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    }
                }
            },
            "DependenceInfo": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "xml": {
                            "attribute": true
                        }
                    },
                    "version": {
                        "type": "string",
                        "xml": {
                            "attribute": true
                        }
                    }
                }
            },
            "Metadata": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    },
                    "engineVersion": {
                        "type": "string"
                    },
                    "author": {
                        "type": "string"
                    },
                    "authorEmail": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "dependencies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DependenceInfo"
                        }
                    }
                }
            },
            "StoredVersionInfo": {
                "type": "object",
                "properties": {
                    "metadata": {
                        "$ref": "#/components/schemas/Metadata"
                    },
                    "source": {
                        "$ref": "#/components/schemas/VersionSourceInfo"
                    },
                    "saveData": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "VersionSourceInfo": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "UNKNOWN",
                            "GITHUB_RELEASE",
                            "OPM_HUB",
                            "MANUAL_PUSH"
                        ]
                    },
                    "versionURL": {
                        "type": "string"
                    },
                    "packageURL": {
                        "type": "string"
                    }
                }
            },
            "ChannelInfo": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    }
                }
            },
            "JobStatus": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "lastStartDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastFinishData": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "launchType": {
                        "type": "string",
                        "enum": [
                            "MANUAL",
                            "SCHEDULED"
                        ]
                    }
                }
            },
            "JobInfo_Public": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/JobStatus_Public"
                    },
                    "settings": {
                        "$ref": "#/components/schemas/JobSettings_Public"
                    }
                }
            },
            "JobSettings_Public": {
                "type": "object",
                "properties": {
                    "schedule": {
                        "type": "boolean"
                    },
                    "cron": {
                        "type": "string"
                    }
                }
            },
            "JobStatus_Public": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "lastStartDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastFinishData": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "launchType": {
                        "type": "string",
                        "enum": [
                            "MANUAL",
                            "SCHEDULED"
                        ]
                    }
                }
            },
            "Response": {
                "type": "object",
                "properties": {
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                }
            }
        }
    }
}